+++ src/lib.rs
@@ -1,10 +1,9 @@
 #![feature(arbitrary_enum_discriminant)]
 
-
-//mod macro_tests;
 mod torpedo_tests;
 mod macro_tests;
 mod cells;
 mod cellcrypto;
 mod custom_crypto;
 mod directories;
+mod misc;